Use sudachipy v0.6.10 or later to fix FTBFS in aarch64 Linux environment
This is a workaround to install the transformer package with the ja component
in aarch64 Linux environment.

The Huggingface Transformer supports several components. Flexeval enables "ja"
component. It depends on sudachipy 0.6.6 or later. My Python 3 runtime selects
sudachipy 0.6.9.

In PyPI, no pre-built binary is provided for sudachipy 0.6.9, thus Pip downloads
the source distribution and attempts to build it.

https://pypi.org/project/SudachiPy/0.6.9/#files

It hits FTBFS of sudachipy 0.6.9 due to invalid entries in Cargo.toml.

https://github.com/WorksApplications/sudachi.rs/issues/290

The issue has been fixed in sudachipy 0.6.10. Let us use the version.